Move format test data to dedicated file.
